Update AddPanel.java
I have updated the Add Panel file to include GUI information, but there seems to be an error with the message that appears in red and the layout of the panel, because all of the text fields are too short to use and things are positioned in a very ugly way. If you get a chance to take a look at this code, let me know what I did wrong, I want to get this fixed first so that we can test the rest of the program.